[refactor] UI: split composable functions into separate file and combine previews
