+++ app/src/main/java/uk/ryanwong/skycatnews/newslist/ui/screen/NewsListScreen.kt
@@ -6,42 +6,25 @@ package uk.ryanwong.skycatnews.newslist.ui.screen
 
 import android.content.res.Configuration.UI_MODE_NIGHT_NO
 import android.content.res.Configuration.UI_MODE_NIGHT_YES
-import androidx.compose.foundation.background
 import androidx.compose.foundation.layout.Box
-import androidx.compose.foundation.layout.PaddingValues
 import androidx.compose.foundation.layout.fillMaxSize
-import androidx.compose.foundation.lazy.LazyColumn
-import androidx.compose.foundation.lazy.itemsIndexed
-import androidx.compose.material.ExperimentalMaterialApi
-import androidx.compose.material.MaterialTheme
 import androidx.compose.material.SnackbarHost
 import androidx.compose.material.SnackbarHostState
-import androidx.compose.material.pullrefresh.PullRefreshIndicator
-import androidx.compose.material.pullrefresh.pullRefresh
-import androidx.compose.material.pullrefresh.rememberPullRefreshState
 import androidx.compose.runtime.Composable
 import androidx.compose.runtime.LaunchedEffect
 import androidx.compose.runtime.remember
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
-import androidx.compose.ui.platform.LocalConfiguration
 import androidx.compose.ui.res.stringResource
-import androidx.compose.ui.semantics.contentDescription
-import androidx.compose.ui.semantics.semantics
 import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.tooling.preview.PreviewParameter
 import uk.ryanwong.skycatnews.R
-import uk.ryanwong.skycatnews.app.ui.component.NoDataScreen
 import uk.ryanwong.skycatnews.app.ui.theme.SkyCatNewsTheme
 import uk.ryanwong.skycatnews.newslist.domain.model.NewsItem
-import uk.ryanwong.skycatnews.newslist.ui.screen.component.LargeStoryHeadline
-import uk.ryanwong.skycatnews.newslist.ui.screen.component.LargeWebLinkHeadline
-import uk.ryanwong.skycatnews.newslist.ui.screen.component.RegularStoryHeadline
-import uk.ryanwong.skycatnews.newslist.ui.screen.component.RegularWebLinkHeadline
 import uk.ryanwong.skycatnews.newslist.ui.screen.previewparameter.NewsListProvider
-import uk.ryanwong.skycatnews.uk.ryanwong.skycatnews.app.ui.theme.getDimension
 import uk.ryanwong.skycatnews.uk.ryanwong.skycatnews.newslist.ui.NewsListUIEvent
 import uk.ryanwong.skycatnews.uk.ryanwong.skycatnews.newslist.ui.NewsListUIState
+import uk.ryanwong.skycatnews.uk.ryanwong.skycatnews.newslist.ui.screen.component.NewsListScreenLayout
 
 @Composable
 fun NewsListScreen(
@@ -81,109 +64,12 @@ fun NewsListScreen(
     }
 }
 
-@OptIn(ExperimentalMaterialApi::class)
-@Composable
-fun NewsListScreenLayout(
-    newsList: List<NewsItem>,
-    isLoading: Boolean,
-    onRefresh: () -> Unit,
-    onStoryItemClicked: (id: Int) -> Unit,
-    onWebLinkItemClicked: (id: Int) -> Unit,
-) {
-    val dimension = LocalConfiguration.current.getDimension()
-    val contentDescriptionNewsList = stringResource(R.string.content_description_news_list)
-    val pullRefreshState = rememberPullRefreshState(refreshing = isLoading, onRefresh = onRefresh)
-
-    Box(
-        modifier = Modifier
-            .fillMaxSize()
-            .background(color = MaterialTheme.colors.background)
-            .pullRefresh(pullRefreshState),
-    ) {
-        LazyColumn(
-            contentPadding = PaddingValues(vertical = dimension.grid_2),
-            modifier = Modifier
-                .fillMaxSize()
-                .semantics { contentDescription = contentDescriptionNewsList }
-        ) {
-            if (newsList.isEmpty()) {
-                if (!isLoading) {
-                    item {
-                        NoDataScreen(modifier = Modifier.fillParentMaxHeight())
-                    }
-                }
-                return@LazyColumn
-            }
-
-            item {
-                val firstItem = newsList.first()
-                when (firstItem) {
-                    is NewsItem.Story -> {
-                        LargeStoryHeadline(
-                            story = firstItem,
-                            onItemClicked = { onStoryItemClicked(firstItem.newsId) },
-                        )
-                    }
-
-                    is NewsItem.WebLink -> {
-                        LargeWebLinkHeadline(
-                            webLink = firstItem,
-                            onItemClicked = { onWebLinkItemClicked(firstItem.newsId) },
-                        )
-                    }
-                }
-            }
-
-            val regularItemsList = newsList.subList(1, newsList.size)
-            itemsIndexed(regularItemsList) { _, newsItem ->
-                when (newsItem) {
-                    is NewsItem.Story -> {
-                        RegularStoryHeadline(
-                            story = newsItem,
-                            onItemClicked = { onStoryItemClicked(newsItem.newsId) },
-                        )
-                    }
-
-                    is NewsItem.WebLink -> {
-                        RegularWebLinkHeadline(
-                            webLink = newsItem,
-                            onItemClicked = { onWebLinkItemClicked(newsItem.newsId) },
-                        )
-                    }
-                }
-            }
-        }
-
-        PullRefreshIndicator(
-            refreshing = isLoading,
-            state = pullRefreshState,
-            modifier = Modifier.align(Alignment.TopCenter)
-        )
-    }
-}
-
 @Preview(
     name = "News List Screen",
     group = "Light",
     showBackground = true,
     uiMode = UI_MODE_NIGHT_NO,
 )
-@Composable
-private fun NewsListScreenPreviewLight(
-    @PreviewParameter(NewsListProvider::class)
-    newsList: List<NewsItem>,
-) {
-    SkyCatNewsTheme {
-        NewsListScreenLayout(
-            newsList = newsList,
-            isLoading = false,
-            onRefresh = { },
-            onStoryItemClicked = {},
-            onWebLinkItemClicked = {},
-        )
-    }
-}
-
 @Preview(
     name = "News List Screen",
     group = "Dark",
@@ -191,17 +77,21 @@ private fun NewsListScreenPreviewLight(
     uiMode = UI_MODE_NIGHT_YES,
 )
 @Composable
-private fun NewsListScreenPreviewDark(
+private fun NewsListScreenPreview(
     @PreviewParameter(NewsListProvider::class)
     newsList: List<NewsItem>,
 ) {
     SkyCatNewsTheme {
-        NewsListScreenLayout(
-            newsList = newsList,
-            isLoading = false,
-            onRefresh = { },
-            onStoryItemClicked = {},
-            onWebLinkItemClicked = {},
+        NewsListScreen(
+            uiState = NewsListUIState(
+                newsList = newsList,
+            ),
+            uiEvent = NewsListUIEvent(
+                onStoryItemClicked = {},
+                onWebLinkItemClicked = {},
+                onRefresh = {},
+                onErrorShown = {},
+            ),
         )
     }
 }
@@ -212,19 +102,6 @@ private fun NewsListScreenPreviewDark(
     showBackground = true,
     uiMode = UI_MODE_NIGHT_NO,
 )
-@Composable
-private fun NewsListScreenNoDataPreviewLight() {
-    SkyCatNewsTheme {
-        NewsListScreenLayout(
-            newsList = emptyList(),
-            isLoading = false,
-            onRefresh = { },
-            onStoryItemClicked = {},
-            onWebLinkItemClicked = {},
-        )
-    }
-}
-
 @Preview(
     name = "News List Screen No Data",
     group = "Dark",
@@ -232,14 +109,18 @@ private fun NewsListScreenNoDataPreviewLight() {
     uiMode = UI_MODE_NIGHT_YES,
 )
 @Composable
-private fun NewsListScreenNoDataPreviewDark() {
+private fun NewsListScreenNoDataPreview() {
     SkyCatNewsTheme {
-        NewsListScreenLayout(
-            newsList = emptyList(),
-            isLoading = false,
-            onRefresh = { },
-            onStoryItemClicked = {},
-            onWebLinkItemClicked = {},
+        NewsListScreen(
+            uiState = NewsListUIState(
+                newsList = emptyList(),
+            ),
+            uiEvent = NewsListUIEvent(
+                onStoryItemClicked = {},
+                onWebLinkItemClicked = {},
+                onRefresh = {},
+                onErrorShown = {},
+            ),
         )
     }
 }
